% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.hmclearn}
\alias{coef.hmclearn}
\title{Extract Model Coefficients}
\usage{
\method{coef}{hmclearn}(x, burnin = NULL, prob = 0.5, ...)
}
\arguments{
\item{x}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{...}{additional arguments to pass to \code{quantile}}

\item{probs}{quantile to extract coefficients}
}
\value{
numeric vector of parameter point estimates
}
\description{
Method for \code{hmclearn} objects created by \code{mh} and \code{hmc} functions.  Extracts the specified quantile of the posterior.
}
