% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagplots.hmclearn}
\alias{diagplots.hmclearn}
\title{Diagnostic plots for \code{hmclearn}}
\usage{
\method{diagplots}{hmclearn}(
  object,
  burnin = NULL,
  plotfun = 2,
  actual.mu = NULL,
  cols = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{plotfun}{integer 1 or 2 indicating which plots to display.  1 shows trace plots.  2 shows a histogram}

\item{actual.mu}{optional numeric vector of true parameter values}

\item{cols}{optional integer index indicating which parameters to display}

\item{...}{currently unused}
}
\description{
Plots histograms of the posterior estimates.  Optionally, displays the 'actual'
values given a simulated dataset.
}
