% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drugs}
\alias{Drugs}
\title{Student Drug Usage Dataset}
\format{
A data frame with 8 rows and 4 variables:
\describe{
  \item{A}{Alcohol usage (Yes/No)}
  \item{C}{Cigarette usage (Yes/No)}
  \item{M}{Marijuana usage (Yes/No)}
  \item{count}{number of responses}
}
}
\source{
Data originally provided by Harry Khamis, Wright State University
}
\usage{
Drugs
}
\description{
Data from a survey of 2276 high school students about drug usage
}
\references{
Agresti, A. (2015). \emph{Foundations of linear and generalized linear models}. John Wiley & Sons.  \url{http://users.stat.ufl.edu/~aa/glm/data/Drugs.dat}
}
\keyword{datasets}
