% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Endometrial}
\alias{Endometrial}
\title{Endometrial Cancer Dataset}
\format{
A data frame with 79 rows and 4 variables:
\describe{
  \item{NV}{Neovasculation risk factor indicator (0=Absent, 1=Present)}
  \item{PI}{Pulsatility index of arteria uterina}
  \item{EH}{Endometrium height}
  \item{HG}{histology of patient (0=Low, 1=High)}
}
}
\source{
\url{http://users.stat.ufl.edu/~aa/glm/data/Endometrial.dat}
}
\usage{
Endometrial
}
\description{
Data from a study about Endometrial Cancer
}
\references{
Agresti, A. (2015). \emph{Foundations of linear and generalized linear models}. John Wiley & Sons.

Copyright 2015, Alan Agresti, Department of Statistics, University of Florida

Heinze, G., & Schemper, M. (2002). \emph{A solution to the problem of separation in logistic regression}. Statistics in medicine, 21(16), 2409-2419.
}
\keyword{datasets}
