% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_glmm_linear.R
\name{g_lmm_posterior}
\alias{g_lmm_posterior}
\title{Gradient of linear Mixed Effects model log posterior}
\usage{
g_lmm_posterior(
  theta,
  y,
  X,
  Z,
  m,
  q = 1,
  A = 10000,
  nueps = 1,
  nulambda = 1,
  Aeps = 25,
  Alambda = 25,
  B = 0.001
)
}
\arguments{
\item{theta}{vector of parameters.  Stored as a single vector in order fixed effect, random effect, log-transformed error, log-transformed diagonal \eqn{\lambda}, and off-diagonal of \code{G} vector \code{a}}

\item{y}{numeric vector for the dependent variable}

\item{X}{numeric design matrix of fixed effect parameters}

\item{Z}{numeric design matrix of random effect parameters}

\item{m}{number of random effect linear parameters}

\item{q}{number of random effects covariance parameters}

\item{A}{hyperprior numeric vector for the random effects off-diagonal \code{a}}

\item{nueps}{hyperprior for the half-t prior of the error parameter \eqn{\nu}}

\item{nulambda}{hyperprior for the half-t prior of the random effects diagonal \eqn{\lambda}}

\item{Aeps}{hyperprior for the half-t prior of the error parameter \eqn{A_\epsilon}}

\item{Alambda}{hyperprior for the half-t prior of the random effects diagonal \eqn{A_\lambda}}

\item{B}{prior for linear predictors is multivariate Normal with mean 0 with diagonal covariance B^-1}
}
\value{
numeric vector of the gradient of the log posterior
}
\description{
Compute the gradient of the log posterior of a linear mixed effects regression model.
Priors are multivariate Normal for the fixed effects
and half-t for the error
}
\details{
The likelihood function for linear mixed effect regression

\deqn{p(y | \beta, u, \sigma_\epsilon^2) &\propto (\sigma_\epsilon^2)^{-n/2} e^{-\frac{1}{2\sigma_\epsilon^2}(y - X\beta - Zu)^T (y - X\beta - Zu)}}

with priors \eqn{\beta \sim N(0, BI)}, \eqn{\sigma_\epsilon \sim half-t(A_\epsilon, nu_\epsilon)}, \eqn{\lambda \sim half-t(A_\lambda, nu_\lambda )}.

The vector \eqn{\lambda} is the diagonal of the covariance \code{G} hyperprior where \eqn{u \sim N(0, G}.  The off-diagonal hyperpriors are stored in a vector \eqn{a \sim N(0, A}.  See Chan, Jeliazkov (2009) for details.

The input parameter vector \code{theta} is of length \code{k}.  The first \code{k-1} parameters are for \eqn{\beta}, and the last parameter is \eqn{\gamma}
}
\references{
Gelman, A. (2006). \emph{Prior distributions for variance parameters in hierarchical models (comment on article by Browne and Draper)}. Bayesian analysis, 1(3), 515-534.

Chan, J. C. C., & Jeliazkov, I. (2009). \emph{MCMC estimation of restricted covariance matrices}. Journal of Computational and Graphical Statistics, 18(2), 457-480.

Betancourt, M., & Girolami, M. (2015). \emph{Hamiltonian Monte Carlo for hierarchical models}. Current trends in Bayesian methodology with applications, 79, 30.
}
