% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_logistic.R
\name{g_logistic_posterior}
\alias{g_logistic_posterior}
\title{Gradient of logistic regression log posterior}
\usage{
g_logistic_posterior(theta, y, X, B = 0.01)
}
\arguments{
\item{theta}{vector of linear predictor parameters}

\item{y}{numeric vector for the dependent variable}

\item{X}{numeric design matrix of independent variables}

\item{B}{prior for linear predictors is multivariate Normal with mean 0 with diagonal covariance B^-1}
}
\value{
numeric vector for the gradient of the log posterior
}
\description{
Compute the gradient of the log posterior of a logistic regression model.
Priors are multivariate Normal for the linear predictors
}
\details{
The likelihood function for logistic regression

\deqn{L(\beta; X, y) = \prod_{i=1}^{n} \left(\frac{1}{1+e^{-X_i\beta}}\right)^{y_i} \left(\frac{e^{-X_i\beta}}{1+e^{-X_i\beta}}\right)^{1-y_i}}

with priors \eqn{\beta \sim N(0, BI)}.

The input parameter vector \code{theta} is of length \code{k}, containing parameter values for \eqn{\beta}
}
