% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_poisson.R
\name{g_poisson_posterior}
\alias{g_poisson_posterior}
\title{Gradient of Poisson regression log posterior}
\usage{
g_poisson_posterior(beta_param, y, X, B = 0.01)
}
\arguments{
\item{y}{numeric vector for the dependent variable}

\item{X}{numeric design matrix of independent variables}

\item{B}{prior for linear predictors is multivariate Normal with mean 0 with diagonal covariance B^-1}

\item{theta}{vector of linear predictor parameters}
}
\value{
numeric vector for the gradient of the log posterior
}
\description{
Compute the gradient of the log posterior of a poisson regression model.
Priors are multivariate Normal for the linear predictors
}
\details{
The likelihood function for poisson regression

\deqn{L(\beta; y, X) = \prod_{i=1}^n \frac{e^{-e^{X_i\beta}}e^{y_iX_i\beta}}{y_i!}}

with priors \eqn{\beta \sim N(0, BI)}.

The input parameter vector \code{theta} is of length \code{k}, containing parameter values for \eqn{\beta}
}
