% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gdat}
\alias{Gdat}
\title{Count of Fresh Gopher Tortoise Shells}
\format{
A data frame with 30 rows and 7 variables:
\describe{
  \item{Site}{name of site}
  \item{year}{years 2004, 2005, 2006}
  \item{shells}{count of shells}
  \item{type}{fresh water}
  \item{Area}{area of the site}
  \item{density}{estimated tortoise density}
  \item{prev}{Seroprevalence to Mycoplasma agassizii}
}
}
\source{
Ozgul, A., Oli, M. K., Bolker, B. M., & Perez-Heydrich, C. (2009). \emph{Upper respiratory tract disease, force of infection, and effects on survival of gopher tortoises}. Ecological Applications, 19(3), 786–798
}
\usage{
Gdat
}
\description{
A dataset containing the count of fresh gopher shells by area.
}
\references{
Fox, G. A., Negrete-Yankelevich, S., & Sosa, V. J. (Eds.). (2015). \emph{Ecological statistics: contemporary theory and application}. Oxford University Press, USA.

Bolker, Ben (2018) GLMM Worked Examples  \url{https://bbolker.github.io/mixedmodels-misc/ecostats_chap.html}
}
\keyword{datasets}
