% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{hmc.fit}
\alias{hmc.fit}
\title{Fit a generic model using Hamiltonian Monte Carlo (HMC)}
\usage{
hmc.fit(
  N,
  theta.init,
  epsilon,
  L,
  logPOSTERIOR,
  glogPOSTERIOR,
  varnames = NULL,
  randlength = FALSE,
  Mdiag = NULL,
  constrain = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{N}{Number of MCMC samples}

\item{theta.init}{Vector of initial values for the parameters}

\item{epsilon}{Step-size parameter for \code{leapfrog}}

\item{L}{Number of \code{leapfrog} steps parameter}

\item{logPOSTERIOR}{Function to calculate and return the log posterior given a vector of values of \code{theta}}

\item{glogPOSTERIOR}{Function to calculate and return the gradient of the log posterior given a vector of values of  \code{theta}}

\item{varnames}{Optional vector of theta parameter names}

\item{randlength}{Logical to determine whether to apply some randomness to the number of leapfrog steps tuning parameter \code{L}}

\item{Mdiag}{Optional vector of the diagonal of the mass matrix \code{M}.  Defaults to unit diagonal.}

\item{constrain}{Optional vector of which parameters in \code{theta} accept positive values only.  Default is that all parameters accept all real numbers}

\item{verbose}{Logical to determine whether to display the progress of the HMC algorithm}

\item{...}{Additional parameters for \code{logPOSTERIOR} and \code{glogPOSTERIOR}}
}
\value{
Object of class \code{hmclearn}
}
\description{
This function runs the HMC algorithm on a generic model provided
the \code{logPOSTERIOR} and gradient \code{glogPOSTERIOR} functions.
All parameters specified in ... are passed to these two functions.
The tuning parameters \code{epsilon} and \code{L} are passed to the
Leapfrog algorithm.
}
\section{Elements for \code{hmclearn} objects}{

\describe{
  \item{\code{N}}{
  Number of MCMC samples
  }
  \item{\code{theta}}{
  List of length \code{N} of the sampled values of \code{theta}
  }
  \item{\code{thetaCombined}}{
  Sampled values in dataframe form
  }
  \item{\code{r}}{
  List of length \code{N} of the sampled momenta
  }
  \item{\code{theta.all}}{
  List of all parameter values of \code{theta} sampled prior to accept/reject step
  }
  \item{\code{r.all}}{
  List of all values of the momenta \code{r} sampled prior to accept/reject
  }
  \item{\code{accept}}{
  Number of accepted proposals.  The ratio \code{accept} / \code{N} is the acceptance rate
  }
  \item{\code{accept_v}}{
  Vector of length \code{N} indicating which samples were accepted
  }
  \item{\code{M_mx}}{
  Mass matrix used in the HMC algorithm
  }
}
}

\section{Available \code{logPOSTERIOR} and \code{glogPOSTERIOR} functions}{

\describe{
  \item{\code{linear_posterior}}{
  Linear regression:  log posterior
  }
  \item{\code{g_linear_posterior}}{
  Linear regression:  gradient of the log posterior
  }
  \item{\code{logistic_posterior}}{
  Logistic regression:  log posterior
  }
  \item{\code{g_logistic_posterior}}{
  Logistic regression:  gradient of the log posterior
  }
  \item{\code{poisson_posterior}}{
  Poisson (count) regression:  log posterior
  }
  \item{\code{g_poisson_posterior}}{
  Poisson (count) regression: gradient of the log posterior
  }
  \item{\code{lmm_posterior}}{
  Linear mixed effects model:  log posterior
  }
  \item{\code{g_lmm_posterior}}{
  Linear mixed effects model:  gradient of the log posterior
  }
  \item{\code{glmm_bin_posterior}}{
  Logistic mixed effects model:  log posterior
  }
  \item{\code{g_glmm_bin_posterior}}{
  Logistic mixed effects model:  gradient of the log posterior
  }
  \item{\code{glmm_poisson_posterior}}{
  Poisson mixed effects model:  log posterior
  }
  \item{\code{g_glmm_poisson_posterior}}{
  Poisson mixed effects model:  gradient of the log posterior
  }
}
}

\references{
\emph{HMC in R} paper

Thomas, S., Li, X., and Tu, W.  2019.  \emph{Hamiltonian Monte Carlo}.  Wiley

Neal, Radford. 2011. \emph{MCMC Using Hamiltonian Dynamics.} In Handbook of Markov Chain Monte Carlo, edited by Steve Brooks, Andrew Gelman, Galin L. Jones, and Xiao-Li Meng, 116–62. Chapman; Hall/CRC.
}
\author{
Samuel Thomas \email{samthoma@iu.edu}, Wanzhu Tu \email{wtu@iu.edu}
}
\keyword{carlo}
\keyword{hamiltonian}
\keyword{monte}
