% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{hmclearn-plots}
\alias{hmclearn-plots}
\alias{mcmc_intervals}
\alias{mcmc_intervals.hmclearn}
\alias{mcmc_areas}
\alias{mcmc_areas.hmclearn}
\alias{mcmc_hist}
\alias{mcmc_hist.hmclearn}
\alias{mcmc_hist_by_chain}
\alias{mcmc_hist_by_chain.hmclearn}
\alias{mcmc_dens}
\alias{mcmc_dens.hmclearn}
\alias{mcmc_scatter}
\alias{mcmc_scatter.hmclearn}
\alias{mcmc_hex}
\alias{mcmc_hex.hmclearn}
\alias{mcmc_pairs}
\alias{mcmc_pairs.hmclearn}
\alias{mcmc_acf}
\alias{mcmc_acf.hmclearn}
\alias{mcmc_acf_bar}
\alias{mcmc_acf_bar.hmclearn}
\alias{mcmc_trace}
\alias{mcmc_trace.hmclearn}
\alias{mcmc_rhat}
\alias{mcmc_rhat.hmclearn}
\alias{mcmc_rhat_hist}
\alias{mcmc_rhat_hist.hmclearn}
\alias{mcmc_neff}
\alias{mcmc_neff.hmclearn}
\alias{mcmc_neff_hist}
\alias{mcmc_neff_hist.hmclearn}
\alias{mcmc_neff_data}
\alias{mcmc_neff_data.hmclearn}
\alias{mcmc_violin}
\alias{mcmc_violin.hmclearn}
\alias{plot.hmclearn}
\title{Plotting for MCMC visualization and diagnostics provided by \code{bayesplot} package}
\usage{
mcmc_intervals(object, ...)

\method{mcmc_intervals}{hmclearn}(object, burnin = NULL, ...)

mcmc_areas(object, ...)

\method{mcmc_areas}{hmclearn}(object, burnin = NULL, ...)

mcmc_hist(object, ...)

\method{mcmc_hist}{hmclearn}(object, burnin = NULL, ...)

mcmc_hist_by_chain(object, ...)

\method{mcmc_hist_by_chain}{hmclearn}(object, burnin = NULL, ...)

mcmc_dens(object, ...)

\method{mcmc_dens}{hmclearn}(object, burnin = NULL, ...)

mcmc_scatter(object, ...)

\method{mcmc_scatter}{hmclearn}(object, burnin = NULL, ...)

mcmc_hex(object, ...)

\method{mcmc_hex}{hmclearn}(object, burnin = NULL, ...)

mcmc_pairs(object, ...)

\method{mcmc_pairs}{hmclearn}(object, burnin = NULL, ...)

mcmc_acf(object, ...)

\method{mcmc_acf}{hmclearn}(object, burnin = NULL, ...)

mcmc_acf_bar(object, ...)

\method{mcmc_acf_bar}{hmclearn}(object, burnin = NULL, ...)

mcmc_trace(object, ...)

\method{mcmc_trace}{hmclearn}(object, burnin = NULL, ...)

mcmc_rhat(object, ...)

\method{mcmc_rhat}{hmclearn}(object, burnin = NULL, ...)

mcmc_rhat_hist(object, ...)

\method{mcmc_rhat_hist}{hmclearn}(object, burnin = NULL, ...)

mcmc_neff(object, ...)

\method{mcmc_neff}{hmclearn}(object, burnin = NULL, lagmax = NULL, ...)

mcmc_neff_hist(object, ...)

\method{mcmc_neff_hist}{hmclearn}(object, burnin = NULL, lagmax = NULL, ...)

mcmc_neff_data(object, ...)

\method{mcmc_neff_data}{hmclearn}(object, burnin = NULL, lagmax = NULL, ...)

mcmc_violin(object, ...)

\method{mcmc_violin}{hmclearn}(object, burnin = NULL, ...)

\method{plot}{hmclearn}(object, burnin = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{...}{optional additional arguments to pass to the \code{bayesplot} functions}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{lagmax}{maximum lag to extract for determining effective sample sizes}
}
\description{
Plots of Rhat statistics, ratios of effective sample size to total sample
size, and autocorrelation of MCMC draws.
}
\section{}{
NA
}

\references{
Gabry, Jonah and Mahr, Tristan (2019).  \emph{bayesplot:  Plotting for Bayesian Models}.  \url{https://mc-stan.org/bayesplot}

Gabry, J., Simpson, D., Vehtari, A., Betancourt, M., and Gelman, A (2019).  \emph{Visualization in Bayesian Workflow}.  Journal of the Royal Statistical Society: Series A. Vol 182.  Issue 2.  p.389-402.

Gelman, A. and Rubin, D. (1992) \emph{Inference from Iterative Simulation Using Multiple Sequences}.  Statistical Science 7(4) 457-472.

Gelman, A., et. al. (2013) \emph{Bayesian Data Analysis}.  Chapman and Hall/CRC.
}
