% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{leapfrog}
\alias{leapfrog}
\title{# nu is a vector
#' @export
qprop_all <- function(theta1, nu) {
  nu <- diag(nu)
  MASS::mvrnorm(1, theta1, nu)
}}
\usage{
leapfrog(
  theta_lf,
  r,
  epsilon,
  logPOSTERIOR,
  glogPOSTERIOR,
  Minv,
  constrain,
  lastSTEP = FALSE,
  ...
)
}
\arguments{
\item{theta_lf}{starting parameter vector}

\item{r}{starting momentum vector}

\item{epsilon}{Step-size parameter for \code{leapfrog}}

\item{logPOSTERIOR}{Function to calculate and return the log posterior given a vector of values of \code{theta}}

\item{glogPOSTERIOR}{Function to calculate and return the gradient of the log posterior given a vector of values of \code{theta}}

\item{Minv}{Inverse Mass matrix}

\item{constrain}{Optional vector of which parameters in \code{theta} accept positive values only.  Default is that all parameters accept all real numbers}

\item{lastSTEP}{Boolean indicating whether to calculate the last half-step of the momentum update}
}
\value{
List containing two elements:  \code{theta.new} the ending value of theta and \code{r.new} the ending value of the momentum
}
\description{
#' @export
qfun_all <- function(theta1, theta2, nu) {
  nu <- diag(nu)
  mvtnorm::dmvnorm(theta1, theta2, nu, log=TRUE)
}
Leapfrog Algorithm for Hamiltonian Monte Carlo
}
\details{
Runs a single iteration of the leapfrog algorithm.  Typically called directly from \code{hmc}
}
\references{
Neal, Radford. 2011. \emph{MCMC Using Hamiltonian Dynamics.} In Handbook of Markov Chain Monte Carlo, edited by Steve Brooks, Andrew Gelman, Galin L. Jones, and Xiao-Li Meng, 116–62. Chapman; Hall/CRC.
}
