% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_linear.R
\name{linear_posterior}
\alias{linear_posterior}
\title{Linear regression log posterior}
\usage{
linear_posterior(theta, y, X, a = 1e-04, b = 1e-04, B = 0.001)
}
\arguments{
\item{theta}{vector of parameters.  All but the last parameter should be linear predictor parameter values.  The last parameter should be the log of the squared error parameter}

\item{y}{numeric vector for the dependent variable}

\item{X}{numeric design matrix of independent variables}

\item{a}{hyperprior for the Inverse Gamma shape parameter}

\item{b}{hyperprior for the Inverse Gamma scale parameter}

\item{B}{inverse of the diagonal of the Covariance hyperprior for the linear predictors}
}
\value{
numeric value for the log posterior
}
\description{
Compute the log posterior of a linear regression model.
Priors are multivariate Normal for the predictor variables
and Inverse Gamma for the error term squared.
}
\details{
The likelihood function for linear regression

\deqn{p(y | X, \beta; \sigma^2) = \frac{1}{(2\pi\sigma^2)^{n/2}}\exp{\left(-\frac{1}{2\sigma^2} (y - X\beta)^T(y-X\beta) \right)}}

with priors \eqn{p(\sigma^2) \sim IG(a, b)} and \eqn{\beta \sim N(0, BI)}.  The variance term is log transformed \eqn{\gamma = \log\sigma^2}

The input parameter vector \code{theta} is of length \code{k}.  The first \code{k-1} parameters are for \eqn{\beta}, and the last parameter is \eqn{\gamma}

Note that the Inverse Gamma prior can be problematic for certain applications with low variance.  See Gelman (2006)
}
\references{
Gelman, A. (2006). \emph{Prior distributions for variance parameters in hierarchical models (comment on article by Browne and Draper)}. Bayesian analysis, 1(3), 515-534.
}
