% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{mh.fit}
\alias{mh.fit}
\title{Fitter function for Metropolis-Hastings (MH)}
\usage{
mh.fit(
  N,
  theta.init,
  qPROP,
  qFUN,
  logPOSTERIOR,
  nu = 0.001,
  varnames = NULL,
  param = list(),
  ...
)
}
\arguments{
\item{N}{Number of MCMC samples}

\item{theta.init}{Vector of initial values for the parameters}

\item{qPROP}{Function to generate proposal}

\item{qFUN}{Probability for proposal function.  First argument is where to evaluate, and second argument is the conditional parameter}

\item{logPOSTERIOR}{Function to calculate and return the log posterior given a vector of values of \code{theta}}

\item{nu}{Single value or vector parameter passed to \code{qPROP} or \code{qFUN} for the proposal density}

\item{varnames}{Optional vector of theta parameter names}

\item{param}{List of additional parameters for \code{logPOSTERIOR}}

\item{...}{Additional parameters for \code{logPOSTERIOR}}
}
\value{
List for \code{mh}
}
\description{
This is the basic computing function for MH and should not be called directly except by experienced users.
}
