% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diag.R
\name{neff}
\alias{neff}
\title{Effective sample size calculation}
\usage{
neff(object, burnin = NULL, lagmax = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{lagmax}{maximum lag to extract for determining effective sample sizes}

\item{...}{currently unused}
}
\value{
numeric vector of effective sample sizes for each parameter
}
\description{
Calculates an estimate of the adjusted MCMC sample size per parameter
adjusted for autocorrelation.
}
\references{
Gelman, A., et. al. (2013) \emph{Bayesian Data Analysis}.  Chapman and Hall/CRC.  Section 11.5
}
