% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot.hmclearn}
\alias{plot.hmclearn}
\title{Plot Histograms of the Posterior Distribution}
\usage{
\method{plot}{hmclearn}(x, burnin = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{...}{optional additional arguments to pass to the \code{bayesplot} functions}
}
\description{
Calls \code{mcmc_hist} from the \code{bayesplot} package to display histograms of the posterior
}
\references{
Gabry, Jonah and Mahr, Tristan (2019).  \emph{bayesplot:  Plotting for Bayesian Models}.  \url{https://mc-stan.org/bayesplot}
}
