% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.hmclearn}
\alias{predict.hmclearn}
\title{Model Predictions for HMC or MH}
\usage{
\method{predict}{hmclearn}(object, y, X, fam = "linear", burnin = NULL, draws = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{y}{vector of the dependent variables used to fit the model}

\item{X}{design matrix, either from fitting the model or new data}

\item{fam}{generalized linear model family.  Currently "linear", "binomial", and "poisson" are supported}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{draws}{Number of simulated values from the posterior conditioned on \code{X}}

\item{...}{additional parameters, currently unsupported}
}
\value{
an object of class \code{hmclearnpred}.  Can be used as an input for \code{pp_check} functions in \code{bayesplot}
}
\description{
\code{predict} generates simulated data from the posterior predictive distribution.
This simulated data can be used for posterior predictive check diagnostics from the \code{bayesplot} package
}
\section{Elements of \code{hmclearnpred} objects}{

\describe{
  \item{\code{y}}{
  Numeric vector of the dependent variable
  }
  \item{\code{ypred}}{
  Matrix of simulated values where each row is a draw from the posterior predictive distribution
  }
  \item{\code{X}}{
  Numeric design matrix
  }
}
}

