% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{qfun}
\alias{qfun}
\title{Multivariate Normal Density of Theta1 | Theta2}
\usage{
qfun(theta1, theta2, nu)
}
\arguments{
\item{theta1}{Vector of current quantiles}

\item{theta2}{Vector for mean parameter}

\item{nu}{Either a single numeric value for the covariance matrix, or a vector for the diagonal}
}
\value{
Multivariate normal density vector
}
\description{
Provided for Random Walk Metropolis algorithm
}
\references{
Alan Genz, Frank Bretz, Tetsuhisa Miwa, Xuefei
Mi, Friedrich Leisch, Fabian Scheipl and Torsten Hothorn (2019).  \emph{mvtnorm: Multivariate Normal and t Distributions}
}
