% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.hmclearn}
\alias{summary.hmclearn}
\title{Summarizing HMC Model Fits}
\usage{
\method{summary}{hmclearn}(object, burnin = NULL, probs = c(0.05, 0.25, 0.5, 0.75, 0.95), ...)
}
\arguments{
\item{object}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{probs}{quantiles to summarize the posterior distribution}

\item{...}{additional arguments to pass to \code{quantile}}
}
\value{
Returns a matrix with posterior quantiles and the posterior scale reduction factor statistic for each parameter.
}
\description{
summary method for class \code{hmclearn}
}
\references{
Gelman, A., et. al. (2013) \emph{Bayesian Data Analysis}.  Chapman and Hall/CRC.

Gelman, A. and Rubin, D. (1992) \emph{Inference from Iterative Simulation Using Multiple Sequences}.  Statistical Science 7(4) 457-472.
}
