% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{hmc.fit}
\alias{hmc.fit}
\title{Fitter function for Hamiltonian Monte Carlo (HMC)}
\usage{
hmc.fit(
  N,
  theta.init,
  epsilon,
  L,
  logPOSTERIOR,
  glogPOSTERIOR,
  varnames = NULL,
  randlength = FALSE,
  Mdiag = NULL,
  constrain = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{N}{Number of MCMC samples}

\item{theta.init}{Vector of initial values for the parameters}

\item{epsilon}{Step-size parameter for \code{leapfrog}}

\item{L}{Number of \code{leapfrog} steps parameter}

\item{logPOSTERIOR}{Function to calculate and return the log posterior given a vector of values of \code{theta}}

\item{glogPOSTERIOR}{Function to calculate and return the gradient of the log posterior given a vector of values of  \code{theta}}

\item{varnames}{Optional vector of theta parameter names}

\item{randlength}{Logical to determine whether to apply some randomness to the number of leapfrog steps tuning parameter \code{L}}

\item{Mdiag}{Optional vector of the diagonal of the mass matrix \code{M}.  Defaults to unit diagonal.}

\item{constrain}{Optional vector of which parameters in \code{theta} accept positive values only.  Default is that all parameters accept all real numbers}

\item{verbose}{Logical to determine whether to display the progress of the HMC algorithm}

\item{...}{additional parameters for \code{logPOSTERIOR} and \code{glogPOSTERIOR}}
}
\value{
List for \code{hmc}
}
\description{
This is the basic computing function for HMC and should not be called directly except by experienced users.
}
