% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{mh}
\alias{mh}
\title{Fit a generic model using Metropolis-Hastings (MH)}
\usage{
mh(N, theta.init, qPROP, qFUN, logPOSTERIOR, nu = 0.001, ...)
}
\arguments{
\item{N}{Number of MCMC samples}

\item{theta.init}{Vector of initial values for the parameters}

\item{qPROP}{Function to generate proposal}

\item{qFUN}{Probability for proposal function.  First argument is where to evaluate, and second argument is the conditional parameter}

\item{logPOSTERIOR}{Function to calculate and return the log posterior given a vector of values of \code{theta}}

\item{nu}{Single value or vector parameter passed to \code{qPROP} or \code{qFUN} for the proposal density}

\item{...}{Additional parameters for \code{logPOSTERIOR}}

\item{varnames}{Optional vector of theta parameter names}
}
\value{
Object of class \code{hmclearn}
}
\description{
This function runs the MH algorithm on a generic model provided
the \code{logPOSTERIOR} function.
All parameters specified in ... are passed to these the posterior function.
The tuning parameters \code{epsilon} and \code{L} are passed to the
Leapfrog algorithm.
}
\section{Elements for \code{hmclearn} objects}{

\describe{
  \item{\code{N}}{
  Number of MCMC samples
  }
  \item{\code{theta}}{
  List of length \code{N} of the sampled values of \code{theta}
  }
  \item{\code{thetaDF}}{
  Sampled values in dataframe form
  }
  \item{\code{r}}{
  NULL for Metropolis-Hastings
  }
  \item{\code{theta.all}}{
  List of all parameter values of \code{theta} sampled prior to accept/reject step
  }
  \item{\code{r.all}}{
  NULL for Metropolis-Hastings
  }
  \item{\code{accept}}{
  Number of accepted proposals.  The ratio \code{accept} / \code{N} is the acceptance rate
  }
  \item{\code{accept_v}}{
  Vector of length \code{N} indicating which samples were accepted
  }
  \item{\code{M_mx}}{
  NULL for Metropolis-Hastings
  }
}
}

\section{Available \code{logPOSTERIOR} functions}{

\describe{
  \item{\code{linear_posterior}}{
  Linear regression:  log posterior
  }
  \item{\code{logistic_posterior}}{
  Logistic regression:  log posterior
  }
  \item{\code{poisson_posterior}}{
  Poisson (count) regression:  log posterior
  }
  \item{\code{lmm_posterior}}{
  Linear mixed effects model:  log posterior
  }
  \item{\code{glmm_bin_posterior}}{
  Logistic mixed effects model:  log posterior
  }
  \item{\code{glmm_poisson_posterior}}{
  Poisson mixed effects model:  log posterior
  }
}
}

\author{
Samuel Thomas \email{samthoma@iu.edu}, Wanzhu Tu \email{wtu@iu.edu}
}
