% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.hmclearn}
\alias{summary.hmclearn}
\title{Summarizing HMC Model Fits}
\usage{
\method{summary}{hmclearn}(x, burnin = NULL, probs = c(0.05, 0.25, 0.5, 0.75, 0.95), ...)
}
\arguments{
\item{x}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}

\item{burnin}{optional numeric parameter for the number of initial MCMC samples to omit from the summary}

\item{probs}{quantiles to summarize the posterior distribution}

\item{...}{additional arguments to pass to \code{quantile}}
}
\description{
summary method for class \code{hmclearn}
}
